# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/users": {
      "get": {
        "operationId": "UserController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all users",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully created"
          },
          "400": {
            "description": "Invalid input"
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "summary": "Create a new user",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me": {
      "get": {
        "operationId": "UserController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return current user profile"
          },
          "401": {
            "description": "Unauthorized"
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UserController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User profile successfully updated"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update current user profile",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UserController_deleteProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User account successfully deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete current user account",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UserController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return user details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "summary": "Get user by ID",
        "tags": [
          "users"
        ]
      }
    },
...
    "/tag": {
      "get": {
        "operationId": "TagController_findAll",
        "parameters": [
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "Number of tags to skip",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "Number of tags to take",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search in name and description",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return list of tags with total count",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tags": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Tag"
                      }
                    },
                    "total": {
                      "type": "number",
                      "example": 100
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all tags with optional filters",
        "tags": [
          "tags"
        ]
      },
      "post": {
        "operationId": "TagController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTagDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tag created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tag"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tag"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new tag (Admin only)",
        "tags": [
          "tags"
        ]
      }
    },
    "/tag/{id}": {
      "get": {
        "operationId": "TagController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return tag details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tag"
                }
              }
            }
          },
          "404": {
            "description": "Tag not found"
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tag"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get tag by ID",
        "tags": [
          "tags"
        ]
      },
      "put": {
        "operationId": "TagController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tag updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tag"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "404": {
            "description": "Tag not found"
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tag"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update a tag (Admin only)",
        "tags": [
          "tags"
        ]
      },
      "delete": {
        "operationId": "TagController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tag deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "404": {
            "description": "Tag not found"
          },
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete a tag (Admin only)",
        "tags": [
          "tags"
        ]
      }
    },
...
  },
  "info": {
    "title": "My API",
    "description": "The API docs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "User ID"
          },
          "email": {
            "type": "string",
            "description": "User email"
          },
          "username": {
            "type": "string",
            "description": "Username"
          },
          "password": {
            "type": "string",
            "description": "User password"
          },
          "fullname": {
            "type": "string",
            "description": "User full name"
          },
          "role": {
            "enum": [
              "admin",
              "user"
            ],
            "type": "string",
            "description": "User role"
          }
        },
        "required": [
          "_id",
          "email",
          "username",
          "password",
          "fullname",
          "role"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email"
          },
          "username": {
            "type": "string",
            "description": "Username"
          },
          "password": {
            "type": "string",
            "description": "User password"
          },
          "fullname": {
            "type": "string",
            "description": "User full name"
          }
        },
        "required": [
          "email",
          "username",
          "password",
          "fullname"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "description": "User password"
          },
          "fullname": {
            "type": "string",
            "description": "User full name"
          }
        }
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
...
      "Vote": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID duy nhất của lượt bình chọn"
          },
          "user": {
            "description": "Người dùng thực hiện bình chọn",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "userId": {
            "type": "string",
            "description": "ID của người dùng thực hiện bình chọn"
          },
          "targetType": {
            "enum": [
              "post",
              "comment"
            ],
            "type": "string",
            "description": "Loại đối tượng được bình chọn (bài viết hoặc bình luận)",
            "example": "post"
          },
          "targetId": {
            "type": "string",
            "description": "ID của đối tượng được bình chọn"
          },
          "voteValue": {
            "enum": [
              1,
              -1
            ],
            "type": "number",
            "description": "Giá trị bình chọn (1 cho upvote, -1 cho downvote)",
            "example": 1
          },
          "post": {
            "description": "Bài viết được bình chọn (nếu có)",
            "allOf": [
              {
                "$ref": "#/components/schemas/Post"
              }
            ]
          },
          "comment": {
            "description": "Bình luận được bình chọn (nếu có)",
            "allOf": [
              {
                "$ref": "#/components/schemas/Comment"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Thời điểm tạo bình chọn"
          }
        },
        "required": [
          "_id",
          "user",
          "userId",
          "targetType",
          "targetId",
          "voteValue",
          "createdAt"
        ]
      },
      "Comment": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID duy nhất của bình luận"
          },
          "post": {
            "description": "Bài viết chứa bình luận này",
            "allOf": [
              {
                "$ref": "#/components/schemas/Post"
              }
            ]
          },
          "postId": {
            "type": "string",
            "description": "ID của bài viết"
          },
          "user": {
            "description": "Tác giả bình luận",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "userId": {
            "type": "string",
            "description": "ID của tác giả"
          },
          "content": {
            "type": "string",
            "description": "Nội dung bình luận"
          },
          "parentComment": {
            "description": "Bình luận cha (nếu là bình luận trả lời)",
            "allOf": [
              {
                "$ref": "#/components/schemas/Comment"
              }
            ]
          },
          "parentCommentId": {
            "type": "string",
            "description": "ID của bình luận cha (nếu có)"
          },
          "replies": {
            "description": "Các bình luận trả lời",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            }
          },
          "votes": {
            "description": "Các lượt bình chọn của bình luận",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vote"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Thời điểm tạo bình luận"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Thời điểm cập nhật bình luận"
          }
        },
        "required": [
          "_id",
          "post",
          "postId",
          "user",
          "userId",
          "content",
          "replies",
          "votes",
          "createdAt",
          "updatedAt"
        ]
      },
      "Post": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID duy nhất của bài viết"
          },
          "title": {
            "type": "string",
            "description": "Tiêu đề bài viết",
            "maxLength": 255
          },
          "content": {
            "type": "string",
            "description": "Nội dung bài viết"
          },
          "user": {
            "description": "Tác giả bài viết",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "userId": {
            "type": "string",
            "description": "ID của tác giả"
          },
          "category": {
            "description": "Danh mục của bài viết",
            "allOf": [
              {
                "$ref": "#/components/schemas/Category"
              }
            ]
          },
          "categoryId": {
            "type": "string",
            "description": "ID của danh mục"
          },
          "tags": {
            "description": "Các thẻ được gắn cho bài viết",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            }
          },
          "comments": {
            "description": "Các bình luận của bài viết",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            }
          },
          "votes": {
            "description": "Các lượt bình chọn của bài viết",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vote"
            }
          },
          "isLocked": {
            "type": "boolean",
            "description": "Trạng thái khóa của bài viết"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Thời điểm tạo bài viết"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Thời điểm cập nhật bài viết"
          }
        },
        "required": [
          "_id",
          "title",
          "content",
          "user",
          "userId",
          "category",
          "categoryId",
          "tags",
          "comments",
          "votes",
          "isLocked",
          "createdAt",
          "updatedAt"
        ]
      },
      "Tag": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID duy nhất của thẻ"
          },
          "tagName": {
            "type": "string",
            "description": "Tên của thẻ",
            "maxLength": 50
          },
          "description": {
            "type": "string",
            "description": "Mô tả của thẻ"
          },
          "posts": {
            "description": "Các bài viết được gắn thẻ này",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Post"
            }
          }
        },
        "required": [
          "_id",
          "tagName",
          "posts"
        ]
      },
...
      "CreateTagDto": {
        "type": "object",
        "properties": {
          "tagName": {
            "type": "string",
            "description": "The name of the tag",
            "maxLength": 50,
            "example": "technology"
          },
          "description": {
            "type": "string",
            "description": "The description of the tag",
            "example": "Posts about technology and innovation"
          }
        },
        "required": [
          "tagName"
        ]
      },
      "UpdateTagDto": {
        "type": "object",
        "properties": {
          "tagName": {
            "type": "string",
            "description": "The name of the tag",
            "maxLength": 50,
            "example": "technology"
          },
          "description": {
            "type": "string",
            "description": "The description of the tag",
            "example": "Posts about technology and innovation"
          }
        }
      },
...
    }
  }
}